requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
celery==5.0.5
djangorestframework==3.12.2
requests==2.25.1
